pyproject.toml
[build-system]
requires = ['setuptools==68.1.2']
build-backend = 'setuptools.build_meta'

[project]
name = 'crudik'
version = '0.0.1'
readme = 'README.md'
requires-python = '>=3.10'
dependencies = [
    "dishka==1.4.2",
    "fastapi[standard]==0.115.6",
    "asyncpg==0.30.0",
    "sqlalchemy==2.0.37",
    "alembic==1.14.0",
    "redis==5.2.1",
    "aiohttp==3.11.13",
    "miniopy-async==1.21.1",
    "argon2-cffi==23.1.0",
    "cryptography==43.0.3",
]

[project.optional-dependencies]
lint = ['ruff==0.9.3', 'mypy==1.14.1']
test = ['pytest==8.3.4', 'pytest-asyncio==0.25.3', 'coverage==7.6.12']
dev = ['crudik[lint, test]']

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::UserWarning",
    'ignore:function ham\(\) is deprecated:DeprecationWarning',
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
include-package-data = true

[tool.mypy]
strict = true
warn_unreachable = true
show_column_numbers = true
show_error_context = true
check_untyped_defs = true
ignore_missing_imports = false
warn_no_return = true

files = ["src/"]

[tool.ruff]
line-length = 120
include = ["pyproject.toml", "src/**/*.py"]
exclude = ["src/crudik/adapters/db/alembic/**/*.py"]

[tool.ruff.lint]
select = ['ALL']

ignore = [
    # Strange and obscure
    'D100',
    'D104',
    'D101',
    'D105',
    'D102',
    'RET504',
    'D103',
    'PLR0913',
    'S101',
    'EM101',
    'TRY003',
    'D107',
    'ARG002',
    'RUF001',
    'PLR0912',
    'C901',
    'PGH003',
    # Does not work correctly
    'S104',
    'ANN401',
]

[[project.authors]]
name = 'lubaskinc0de'
email = 'lubaskincorporation@gmail.com'

[project.scripts]
crudik = "crudik.bootstrap.cli:main"
